requirements.txt
pypdf
PyPDF2
PyMuPDF
google-generativeai
Pillow
python-dotenv
pydantic
spacy
regex
groq
numpy
transformers
langchain
langchain_community
langchain_groq
pymongo
langchain
python-multipart
langchain-community
langchain-text-splitters
streamlit_chat
sentence_transformers
langchain_huggingface
faiss-cpu
fastapi
uvicorn
requests
chromadb
pinecone-client
pinecone-text
pinecone-notebooks
pymango